Use BigDecimal for monetary values and exchange rates

+++ src/main/java/com/example/fx/model/dto/ConversionHistoryResponse.java
@@ -3,6 +3,7 @@ package com.example.fx.model.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -12,8 +13,8 @@ public class ConversionHistoryResponse {
     private String transactionId;
     private String fromCurrency;
     private String toCurrency;
-    private double originalAmount;
-    private double convertedAmount;
-    private double rate;
+    private BigDecimal originalAmount;
+    private BigDecimal convertedAmount;
+    private BigDecimal rate;
     private LocalDateTime transactionDate;
 }
