+++ src/main/java/com/example/fx/model/dto/ConversionResponse.java
@@ -3,6 +3,8 @@ package com.example.fx.model.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ConversionResponse {
@@ -10,7 +12,7 @@ public class ConversionResponse {
     private String transactionId;
     private String from;
     private String to;
-    private double originalAmount;
-    private double convertedAmount;
-    private double rate;
+    private BigDecimal originalAmount;
+    private BigDecimal convertedAmount;
+    private BigDecimal rate;
 }
