+++ src/main/java/com/example/fx/model/dto/ExchangeRateResponse.java
@@ -3,12 +3,13 @@ package com.example.fx.model.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ExchangeRateResponse {
 
     private String from;
     private String to;
-    private double rate;
-
+    private BigDecimal rate;
 }
